Pacify clippy

+++ benches/curve_jacobian.rs
@@ -128,8 +128,8 @@ fn criterion_benchmark(c: &mut Criterion) {
     for &batch_size in BATCH_SIZES.iter() {
         let ct_name = ct_batch_str.clone() + &batch_size.to_string();
         let vt_name = vt_batch_str.clone() + &batch_size.to_string();
-        let jacobian_points = vec![JacobianPoint::random(&mut rng); batch_size as usize];
-        let scalars = vec![Scalar::random(&mut rng).to_bytes(); batch_size as usize];
+        let jacobian_points = vec![JacobianPoint::random(&mut rng); batch_size];
+        let scalars = vec![Scalar::random(&mut rng).to_bytes(); batch_size];
         c.bench_function(&ct_name, |bench| {
             bench.iter(|| {
                 JacobianPoint::multiply_many(black_box(&jacobian_points), black_box(&scalars))
